% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{batch_suspend_user}
\alias{batch_suspend_user}
\title{Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account}
\usage{
batch_suspend_user(AccountId, UserIdList)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserIdList}{[required] The request containing the user IDs to suspend.}
}
\description{
Suspends up to 50 users from a \code{Team} or \code{EnterpriseLWA} Amazon Chime account. For more information about different account types, see \href{http://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts} in the \emph{Amazon Chime Administration Guide}.
}
\details{
Users suspended from a \code{Team} account are dissociated from the account, but they can continue to use Amazon Chime as free users. To remove the suspension from suspended \code{Team} account users, invite them to the \code{Team} account again. You can use the InviteUsers action to do so.

Users suspended from an \code{EnterpriseLWA} account are immediately signed out of Amazon Chime and are no longer able to sign in. To remove the suspension from suspended \code{EnterpriseLWA} account users, use the BatchUnsuspendUser action.

To sign out users without suspending them, use the LogoutUser action.
}
\section{Accepted Parameters}{
\preformatted{batch_suspend_user(
  AccountId = "string",
  UserIdList = list(
    "string"
  )
)
}
}

