% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{batch_unsuspend_user}
\alias{batch_unsuspend_user}
\title{Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime \code{EnterpriseLWA} account}
\usage{
batch_unsuspend_user(
  AccountId = "string",
  UserIdList = list(
    "string"
  )
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserIdList}{[required] The request containing the user IDs to unsuspend.}
}
\description{
Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime \code{EnterpriseLWA} account. Only users on \code{EnterpriseLWA} accounts can be unsuspended using this action. For more information about different account types, see \href{http://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts} in the \emph{Amazon Chime Administration Guide}.
}
\details{
Previously suspended users who are unsuspended using this action are returned to \code{Registered} status. Users who are not previously suspended are ignored.
}
\examples{

}
