% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{batch_update_user}
\alias{batch_update_user}
\title{Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account}
\usage{
batch_update_user(
  AccountId = "string",
  UpdateUserRequestItems = list(
    list(
      UserId = "string",
      LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial"
    )
  )
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UpdateUserRequestItems}{[required] The request containing the user IDs and details to update.}
}
\description{
Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account. Currently, only \code{LicenseType} updates are supported for this action.
}
\examples{

}
