% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{create_account}
\alias{create_account}
\title{Creates an Amazon Chime account under the administrator's AWS account}
\usage{
create_account(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the Amazon Chime account.}
}
\description{
Creates an Amazon Chime account under the administrator's AWS account. Only \code{Team} account types are currently supported for this action. For more information about different account types, see \href{http://docs.aws.amazon.com/chime/latest/ag/manage-chime-account.html}{Managing Your Amazon Chime Accounts} in the \emph{Amazon Chime Administration Guide}.
}
\examples{

}
