% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{delete_account}
\alias{delete_account}
\title{Deletes the specified Amazon Chime account}
\usage{
delete_account(
  AccountId = "string"
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Deletes the specified Amazon Chime account. You must suspend all users before deleting a \code{Team} account. You can use the BatchSuspendUser action to do so.
}
\details{
For \code{EnterpriseLWA} and \code{EnterpriseAD} accounts, you must release the claimed domains for your Amazon Chime account before deletion. As soon as you release the domain, all users under that account are suspended.

Deleted accounts appear in your \code{Disabled} accounts list for 90 days. To restore a deleted account from your \code{Disabled} accounts list, you must contact AWS Support.

After 90 days, deleted accounts are permanently removed from your \code{Disabled} accounts list.
}
\examples{

}
