% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{get_account}
\alias{get_account}
\title{Retrieves details for the specified Amazon Chime account, such as account type and supported licenses}
\usage{
get_account(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Retrieves details for the specified Amazon Chime account, such as account type and supported licenses.
}
\section{Accepted Parameters}{
\preformatted{get_account(
  AccountId = "string"
)
}
}

