% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{get_account_settings}
\alias{get_account_settings}
\title{Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings}
\usage{
get_account_settings(
  AccountId = "string"
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}
}
\description{
Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings. For more information about these settings, see \href{http://docs.aws.amazon.com/chime/latest/ag/policies.html}{Use the Policies Page} in the \emph{Amazon Chime Administration Guide}.
}
\examples{

}
