% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{logout_user}
\alias{logout_user}
\title{Logs out the specified user from all of the devices they are currently logged into}
\usage{
logout_user(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\description{
Logs out the specified user from all of the devices they are currently logged into.
}
\section{Accepted Parameters}{
\preformatted{logout_user(
  AccountId = "string",
  UserId = "string"
)
}
}

