% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_package.R
\docType{package}
\name{paws.chime-package}
\alias{paws.chime}
\alias{paws.chime-package}
\title{paws.chime: Amazon Chime}
\description{
The Amazon Chime API (application programming interface) is
designed for administrators to use to perform key tasks, such as
creating and managing Amazon Chime accounts and users. This guide
provides detailed information about the Amazon Chime API, including
operations, types, inputs and outputs, and error codes.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
