% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{reset_personal_pin}
\alias{reset_personal_pin}
\title{Resets the personal meeting PIN for the specified user on an Amazon Chime account}
\usage{
reset_personal_pin(
  AccountId = "string",
  UserId = "string"
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\description{
Resets the personal meeting PIN for the specified user on an Amazon Chime account. Returns the User object with the updated personal meeting PIN.
}
\examples{

}
