% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{update_account}
\alias{update_account}
\title{Updates account details for the specified Amazon Chime account}
\usage{
update_account(
  AccountId = "string",
  Name = "string"
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{Name}{The new name for the specified Amazon Chime account.}
}
\description{
Updates account details for the specified Amazon Chime account. Currently, only account name updates are supported for this action.
}
\examples{

}
