% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_operations.R
\name{update_user}
\alias{update_user}
\title{Updates user details for a specified user ID}
\usage{
update_user(
  AccountId = "string",
  UserId = "string",
  LicenseType = "Basic"|"Plus"|"Pro"|"ProTrial"
)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}

\item{LicenseType}{The user license type to update. This must be a supported license type for the Amazon Chime account that the user belongs to.}
}
\description{
Updates user details for a specified user ID. Currently, only \code{LicenseType} updates are supported for this action.
}
\examples{

}
