% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.chime_package.R
\docType{package}
\name{paws.chime-package}
\alias{paws.chime}
\alias{paws.chime-package}
\title{paws.chime: Amazon Chime}
\description{
Amazon Chime is a communications service that
    transforms online meetings with a secure, easy-to-use application that
    you can trust. Amazon Chime works seamlessly across your devices so
    that you can stay connected. You can use Amazon Chime for online
    meetings, video conferencing, calls, and chat. You can also share
    content, both inside and outside your organization. Amazon Chime frees
    you to work productively from anywhere.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
